ported confidence intervals for mean
The function confmean computes the boundaries for upper and lower confidence limits.
It uses the function resultant_vector_length, which was the matlab function r. I gave
it a more meaningful name resultant_vector_length. We could still define a synonym "r"
for it. The function mean_ci is a wrapper for mean and confmean, and computes the mean
with the upper and lower confidence intervals.